Nimrat

+++ src/use_case/delete/DeleteInteractor.java
@@ -2,18 +2,32 @@ package use_case.delete;
 
 import java.util.UUID;
 
-public class DeleteInteractor implements DeleteInputBoundary{
+public class DeleteInteractor implements DeleteInputBoundary {
     final DeleteOutputBoundary deletePresenter;
 
-    public DeleteInteractor(DeleteOutputBoundary deletePresenter) {
+    final DeleteResponseDataAccessInterface responseDataAccessInterface;
+
+    final DeleteUserDataAccessInterface userDataAccessInterface;
+
+    public DeleteInteractor(DeleteOutputBoundary deletePresenter, DeleteResponseDataAccessInterface responseDataAccessInterface, DeleteUserDataAccessInterface userDataAccessInterface) {
         this.deletePresenter = deletePresenter;
+        this.responseDataAccessInterface = responseDataAccessInterface;
+        this.userDataAccessInterface = userDataAccessInterface;
     }
 
     @Override
     public void execute(DeleteInputData deleteInputData) {
         UUID responseId = deleteInputData.getResponseId();
-        DeleteOutputData deleteOutputData = new DeleteOutputData(responseId);
-        deletePresenter.prepareSuccessView(deleteOutputData);
-    }
+        if (responseDataAccessInterface.existsById(responseId) && userDataAccessInterface.responseExistsById(responseId)) {
+            responseDataAccessInterface.deleteResponse(responseId);
+            userDataAccessInterface.deleteResponse(responseId);
 
+            DeleteOutputData deleteOutputData = new DeleteOutputData(responseId);
+            deletePresenter.prepareSuccessView(deleteOutputData);
+        } else {
+            deletePresenter.prepareFailView("Response does not exist");
+
+        }
+
+    }
 }
